backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=24.1.0",
    "fastapi>=0.115.6",
    "gitpython>=3.1.43",
    "pydantic-settings>=2.7.0",
    "python-multipart>=0.0.19",
    "uvicorn[standard]>=0.32.1",
    "watchfiles>=1.0.4",
]

[dev-dependencies]
httpx = ">=0.28.1"
pytest = ">=8.3.4"
ruff = ">=0.9.6"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["B008", "E501"]
